move zkShuffle to jssdk
